Conflicts:
	openair1/PHY/LTE_TRANSPORT/dci_tools.c
	openair1/PHY/LTE_TRANSPORT/dlsch_demodulation.c
	openair1/PHY/LTE_TRANSPORT/dlsch_modulation.c
	openair1/PHY/LTE_TRANSPORT/proto.h
	openair1/SIMULATION/LTE_PHY/dlsim.c

+++ cmake_targets/autotests/tools/configure_cots_huaweiE398_ue.py
@@ -143,6 +143,7 @@ def send_command (cmd, response, timeout):
         error = error + ' In function: ' + sys._getframe().f_code.co_name + ': *** Caught exception: '  + str(e.__class__) + " : " + str( e)
         error = error + traceback.format_exc()
         print error
+        time.sleep(1)
         
 
 def start_ue () :
@@ -180,11 +181,16 @@ def start_ue () :
    thread_ppp.join()
 
 def stop_ue():
-   timeout=60
-   os.system('killall wvdial')
-   send_command('AT', 'OK' , timeout)
-   send_command('AT+CGATT=0' , 'OK|ERROR' , timeout)
-   #send_command('AT+CFUN=4' , 'OK' , timeout)
+  stringIdBandrich='Huawei Technologies Co., Ltd. E398 LTE/UMTS/GSM Modem/Networkcard'
+  status, out = commands.getstatusoutput('lsusb | grep -i \'' + stringIdBandrich + '\'')
+  if (out == '') :
+     print "Huawei E398 Adapter not found. Exiting now..."
+     sys.exit()
+  timeout=60
+  os.system('killall wvdial')
+  send_command('AT', 'OK' , timeout)
+  send_command('AT+CGATT=0' , 'OK|ERROR' , timeout)
+  #send_command('AT+CFUN=4' , 'OK' , timeout)
 
 
 #reset the USB BUS of Bandrich UE
@@ -192,7 +198,7 @@ def reset_ue():
   stringIdBandrich='Huawei Technologies Co., Ltd. E398 LTE/UMTS/GSM Modem/Networkcard'
   status, out = commands.getstatusoutput('lsusb | grep -i \'' + stringIdBandrich + '\'')
   if (out == '') :
-     print "Bandrich 4G LTE Adapter not found. Exiting now..."
+     print "Huawei E398 Adapter not found. Exiting now..."
      sys.exit()
   p=re.compile('Bus\s*(\w+)\s*Device\s*(\w+):\s*ID\s*(\w+):(\w+)')
   res=p.findall(out)
@@ -206,6 +212,7 @@ def reset_ue():
   os.system(cmd + " ; sleep 15" )
   cmd = "sudo sh -c \"echo 1 > " + usb_dir + "/authorized\""
   os.system(cmd + " ; sleep 30" )
+  stop_ue()
 
 i=1
 gw='192.172.0.1'
@@ -220,6 +227,7 @@ while i <  len(sys.argv):
         print 'Using Serial port : ' + serial_port  
         stop_ue()
     elif arg == '--reset-ue' :
+        find_open_port()
         reset_ue()
     elif arg == '-gw' :
         gw = sys.argv[i+1]
